+++ ChangeLog.md
@@ -1,4 +1,15 @@
+### 0.3.0 / 2018-05-16
+
+* Added more options to the CLI
+* Added `#run` and `#open` methods to the Runner
+
+### 0.2.0 / 2018-05-08
+
+* Added `CypressRails::Runner` class that runs Cypress with the specified command inside the
+    `CypressRails::Server.start` block
+* Runner returns the status code of the Cypress process
+
 ### 0.1.0 / 2018-05-04
 
-* Initial release:
-
+* Initial release
+* Added `CypressRails::Server` class that starts the application to test under the specified address